RSA Public-Key Cryptosystem
In this project, block cipher is executed in ECB mode, RSA encryption algorithm is used. Key pairs are generated in the “KeyGeneration” class. Operations with p, q, e, n, d values are done here. Public key and Private key values and whether they are prime can be tested here. Key pairs are generated in the "AllKeys" class. In the "RSA_Algorithm" class, RSA encryption and decryption algorithms are used. ASCII and string conversion methods are included in this class. Ciphertext stealing algorithm is also implemented in this class. A simple GUI has been developed in the "Main" class.